Add support for 2-10-10-10 screencopy buffers
While Cairo has a CAIRO_FORMAT_RGB30 format that would be compatible
with WL_SHM_FORMAT_XRGB2101010, it has no direct way to represent
WL_SHM_FORMAT_ARGB2101010 ; so we convert all the 2-10-10-10 formats
to a floating point representation with almost no loss of accuracy.

The RGBA128F/RGB96F formats are only available with Cairo >= 1.17.2,
which is technically a development snapshot, not an official release.