Read XDG_PICTURES_DIR from user-dirs.dirs file
According to the spec, XDG_PICTURES_DIR is an entry in $XDG_CONFIG_HOME/user-dirs.dirs.
This file can be sourced which will set the env var, however it may not always be.
Hence we should read the path from the file rather than trying to get it from the env.

spec: https://www.freedesktop.org/wiki/Software/xdg-user-dirs/